Fix remaining V3 compilation issues
- Disable submodules checkout in `.github/workflows/codeql.yml`
- Cast `event.data.key.keyCode` to `(int)` in switch statement in `main.cpp` to prevent `-Werror=switch` errors for unhandled enums.
- Use `uint32_t` for `LCD_GetSize` and assign to `int` variables in `calc.cpp` to prevent `invalid conversion` errors.

+++ src/main.cpp
@@ -798,7 +798,7 @@ void main2(){
 						switch(event.key.keysym.sym) {
 				#else
 					if (event.type==EVENT_KEY&&event.data.key.direction==KEY_PRESSED){
-						switch(event.data.key.keyCode) {
+						switch((int)event.data.key.keyCode) {
 				#endif
 							case K_QUIT:
 							#ifdef K_QUIT2
